backend/requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
psycopg2-binary==2.9.5
python-dotenv==0.21.0
Flask-Cors==3.0.10  # Important for allowing frontend to talk to backend